IFS-10081: unit test fixed

+++ ifs-web-service/ifs-competition-mgt-service/src/main/java/org/innovateuk/ifs/management/competition/setup/CompetitionSetupTermsAndConditionsController.java
@@ -344,7 +344,8 @@ public class CompetitionSetupTermsAndConditionsController {
 
     private CompetitionResource deleteDataInTermsSetupswitch(boolean isThirdPartyProcurement, boolean isProcurement, CompetitionResource competition, BindingResult bindingResult) {
         if (competition.getCompetitionTerms() != null)  {
-            if (isProcurement && competition.getCompetitionThirdPartyConfigResource().getTermsAndConditionsLabel() != null) {
+            if (isProcurement && competition.getCompetitionThirdPartyConfigResource() != null
+                    && competition.getCompetitionThirdPartyConfigResource().getTermsAndConditionsLabel() != null) {
                  competitionSetupRestService.deleteCompetitionThirdPartyConfigData(competition.getId());
             }
             if (isThirdPartyProcurement && bindingResult.hasErrors()) {
